OSL-210: minor cleanup
