+++ src/snowplow/testUtils.ts
@@ -1,0 +1,25 @@
+import fetch from 'node-fetch';
+import { config } from '../config';
+
+async function snowplowRequest(path: string, post = false): Promise<any> {
+  const response = await fetch(`http://${config.snowplow.endpoint}${path}`, {
+    method: post ? 'POST' : 'GET',
+  });
+  return await response.json();
+}
+
+export async function resetSnowplowEvents(): Promise<void> {
+  await snowplowRequest('/micro/reset', true);
+}
+
+export async function getAllSnowplowEvents(): Promise<{ [key: string]: any }> {
+  return snowplowRequest('/micro/all');
+}
+
+export async function getGoodSnowplowEvents(): Promise<{ [key: string]: any }> {
+  return snowplowRequest('/micro/good');
+}
+
+export function parseSnowplowData(data: string): { [key: string]: any } {
+  return JSON.parse(Buffer.from(data, 'base64').toString());
+}
